+++ src/App.css
@@ -1,14 +1,14 @@
 
 @font-face {
   font-family: Machina;
-  src: url(./modules/fonts/neue-machina-free-for-personal-use/Neue\ Machina\ -\ Free\ for\ Personal\ Use/NeueMachina-Regular.otf);
+  src: url(./assets/fonts/neue-machina-free-for-personal-use/Neue\ Machina\ -\ Free\ for\ Personal\ Use/NeueMachina-Regular.otf);
 }
 
 @import url('https://fonts.cdnfonts.com/css/script-mt');
 
 @font-face {
   font-family: V;
-  src: url(./modules/fonts/script-mt-bold-font/SCRIPTBL.TTF);
+  src: url(./assets/fonts/script-mt-bold-font/SCRIPTBL.TTF);
 }
 
 html, body { height: 100%; }
