+++ src/components/index.jsx
@@ -6,3 +6,4 @@ export {default as UserProfile} from './UserProfile';
 export {default as Notification} from './Notification';
 export {default as Navbar} from './Navbar';
 export {default as Sidebar} from './Sidebar';
+export {default as NotFound} from './NotFound'
