+++ src/modules/admin/components/index.jsx
@@ -1,0 +1,1 @@
+export {default as AdminLogin} from './AdminLogin'
