--- src/modules/admin/pages/index.jsx
@@ -1,1 +1,0 @@
-export {default as AdminLogin} from './AdminLogin'
