+++ src/modules/landingPage/assets/css/Sign.module.css
@@ -14,7 +14,7 @@
   position: absolute;
   top: -198px;
   width: 6400px;
-  height: 198px;
+  height: 198px; 
   animation: wave 7s cubic-bezier( 0.36, 0.45, 0.63, 0.53) infinite;
   transform: translate3d(0, 0, 0);
 }
@@ -40,5 +40,11 @@
   }
   50% {
     transform: translate3d(0,5px,0);
+  }
+}
+
+@media only screen and (max-width: 600px) {
+  .wave{
+    width:max-content;
   }
 }
