+++ src/modules/landingPage/components/index.jsx
@@ -3,4 +3,3 @@ export {default as CourseRegistration} from './CourseRegistration'
 export {default as Header} from './Header'
 export {default as Login} from './Login'
 export {default as Register} from './Register'
-export {default as NotFound} from './NotFound'
